+++ docs/changes.rst
@@ -9,6 +9,7 @@ Not yet released.
 * Added ``ignore-all-checks`` flag to ignore all quality checks on a string.
 * Avoid :ref:`addon-weblate.generate.pseudolocale` add-on to trigger failing checks.
 * Added support for :ref:`vcs-gitea`.
+* Added Linux style language code to :ref:`component-language_code_style`.
 
 `All changes in detail <https://github.com/WeblateOrg/weblate/milestone/77?closed=1>`__.